# packages/contracts/generated/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateMemoryItemBody": {
        "properties": {
          "assertion_text": {
            "minLength": 1,
            "title": "Assertion Text",
            "type": "string"
          },
          "space_id": {
            "format": "uuid",
            "title": "Space Id",
            "type": "string"
          },
          "title": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "space_id",
          "title",
          "assertion_text"
        ],
        "title": "CreateMemoryItemBody",
        "type": "object"
      },
      "CreateProposalBody": {
        "properties": {
          "base_revision_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Revision Id"
          },
          "patch_json": {
            "additionalProperties": true,
            "title": "Patch Json",
            "type": "object"
          },
          "space_id": {
            "format": "uuid",
            "title": "Space Id",
            "type": "string"
          },
          "target_id": {
            "format": "uuid",
            "title": "Target Id",
            "type": "string"
          },
          "target_type": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Target Type",
            "type": "string"
          }
        },
        "required": [
          "space_id",
          "target_type",
          "target_id"
        ],
        "title": "CreateProposalBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Canonical domain API. Adds documents, memory, context, sources, jobs, and ingestion routes.",
    "title": "Memdot Core API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/memory/items": {
      "post": {
        "operationId": "create_memory_item_api_v1_memory_items_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMemoryItemBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Memory Item",
        "tags": [
          "memory"
        ]
      }
    },
    "/api/v1/memory/items/{memory_item_id}": {
      "get": {
        "operationId": "get_memory_item_api_v1_memory_items__memory_item_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "memory_item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Memory Item Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Memory Item",
        "tags": [
          "memory"
        ]
      }
    },
    "/api/v1/memory/proposals": {
      "post": {
        "operationId": "create_proposal_api_v1_memory_proposals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProposalBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Proposal",
        "tags": [
          "memory"
        ]
      }
    },
    "/api/v1/memory/proposals/{proposal_id}/approve": {
      "post": {
        "operationId": "approve_proposal_api_v1_memory_proposals__proposal_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Proposal",
        "tags": [
          "memory"
        ]
      }
    },
    "/api/v1/memory/proposals/{proposal_id}/reject": {
      "post": {
        "operationId": "reject_proposal_api_v1_memory_proposals__proposal_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Proposal",
        "tags": [
          "memory"
        ]
      }
    },
...
  }
}